Extract task dependency controller slice
