+++ apps/core-api/src/tasks/tasks.controller.ts
@@ -36,7 +36,7 @@ import { DomainService } from '../common/domain.service';
 import { CurrentRequest } from '../common/current-request';
 import type { AppRequest } from '../common/types';
 import { assertValidDateRange, normalizeDateOnlyField, parseTaskDateQuery, toDateOnlyDate } from '../common/date-validation';
-import { Prisma, Priority, ProjectRole, TaskStatus, TaskType, DependencyType, CustomFieldType } from '@prisma/client';
+import { Prisma, Priority, ProjectRole, TaskStatus, TaskType, CustomFieldType } from '@prisma/client';
 import {
   completeTaskLifecycle,
   DomainConflictError,
@@ -44,7 +44,6 @@ import {
   normalizeProjectViewState,
   type ProjectViewState,
 } from '@atlaspm/domain';
-import { SubtaskService } from './subtask.service';
 import { SearchService } from '../search/search.service';
 import { createTaskLifecycleUnitOfWorkFromTx } from './task-lifecycle-prisma.adapter';
 import {
@@ -416,48 +415,6 @@ class PatchDescriptionDto {
   expectedVersion!: number;
 }
 
-class CreateSubtaskDto {
-  @IsString()
-  title!: string;
-
-  @IsOptional()
-  @IsString()
-  description?: string;
-
-  @IsOptional()
-  @IsEnum(TaskStatus)
-  status?: TaskStatus;
-
-  @IsOptional()
-  @IsEnum(Priority)
-  priority?: Priority;
-
-  @IsOptional()
-  @IsString()
-  assigneeUserId?: string;
-
-  @IsOptional()
-  @IsISO8601()
-  startAt?: string;
-
-  @IsOptional()
-  @IsISO8601()
-  dueAt?: string;
-
-  @IsOptional()
-  @IsArray()
-  tags?: string[];
-}
-
-class AddDependencyDto {
-  @IsUUID()
-  dependsOnId!: string;
-
-  @IsOptional()
-  @IsEnum(DependencyType)
-  type?: DependencyType;
-}
-
 const MAX_DESCRIPTION_DOC_BYTES = 200_000;
 const MAX_DESCRIPTION_TEXT_LENGTH = 20_000;
 const TIMELINE_LANE_ORDER_BASE_LIMIT = 500;
@@ -537,7 +494,6 @@ export class TasksController {
   constructor(
     @Inject(PrismaService) private readonly prisma: PrismaService,
     @Inject(DomainService) private readonly domain: DomainService,
-    @Inject(SubtaskService) private readonly subtaskService: SubtaskService,
     @Inject(SearchService) private readonly searchService: SearchService,
     @Inject(NotificationsService) private readonly notifications: NotificationsService,
     @Inject(TaskMentionsService) private readonly mentions: TaskMentionsService,
@@ -2056,100 +2012,6 @@ export class TasksController {
       });
       return { task: updated, sectionTasks };
     });
-  }
-
-  // Subtask endpoints
-  @Post('tasks/:id/subtasks')
-  async createSubtask(@Param('id') parentId: string, @Body() body: CreateSubtaskDto, @CurrentRequest() req: AppRequest) {
-    const parentTask = await this.prisma.task.findFirstOrThrow({ where: { id: parentId, deletedAt: null } });
-    await this.domain.requireProjectRole(parentTask.projectId, req.user.sub, ProjectRole.MEMBER);
-    if (parentTask.parentId) {
-      throw new BadRequestException('Nested subtasks are not supported');
-    }
-    assertValidDateRange(body.startAt, body.dueAt);
-    const startAt = toDateOnlyDate(body.startAt) ?? null;
-    const dueAt = toDateOnlyDate(body.dueAt) ?? null;
-
-    const topTask = await this.prisma.task.findFirst({
-      where: { projectId: parentTask.projectId, sectionId: parentTask.sectionId, deletedAt: null },
-      orderBy: { position: 'asc' },
-    });
-    const position = (topTask?.position ?? 1000) - 1000;
-
-    const taskData = {
-      ...body,
-      projectId: parentTask.projectId,
-      sectionId: parentTask.sectionId,
-      position,
-      startAt,
-      dueAt,
-    };
-    return this.subtaskService.createSubtask(parentId, taskData);
-  }
-
-  @Get('tasks/:id/subtasks')
-  async getSubtasks(@Param('id') taskId: string, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.VIEWER);
-    return this.subtaskService.getSubtasks(taskId);
-  }
-
-  @Get('tasks/:id/subtasks/tree')
-  async getSubtaskTree(@Param('id') taskId: string, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.VIEWER);
-    return this.subtaskService.getSubtaskTree(taskId);
-  }
-
-  @Get('tasks/:id/breadcrumbs')
-  async getBreadcrumbs(@Param('id') taskId: string, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.VIEWER);
-    return this.subtaskService.getBreadcrumbPath(taskId);
-  }
-
-  // Dependency endpoints
-  @Post('tasks/:id/dependencies')
-  async addDependency(@Param('id') taskId: string, @Body() body: AddDependencyDto, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.MEMBER);
-    return this.subtaskService.addDependency(taskId, body.dependsOnId, body.type, req.user.sub, req.correlationId);
-  }
-
-  @Delete('tasks/:id/dependencies/:dependsOnId')
-  async removeDependency(@Param('id') taskId: string, @Param('dependsOnId') dependsOnId: string, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.MEMBER);
-    await this.subtaskService.removeDependencyWithAudit(taskId, dependsOnId, req.user.sub, req.correlationId);
-    return { success: true };
-  }
-
-  @Get('tasks/:id/dependencies')
-  async getDependencies(@Param('id') taskId: string, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.VIEWER);
-    return this.subtaskService.getDependencies(taskId);
-  }
-
-  @Get('tasks/:id/dependents')
-  async getDependents(@Param('id') taskId: string, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.VIEWER);
-    return this.subtaskService.getDependents(taskId);
-  }
-
-  @Get('tasks/:id/blocked')
-  async isBlocked(@Param('id') taskId: string, @CurrentRequest() req: AppRequest) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.VIEWER);
-    const blocked = await this.subtaskService.isBlocked(taskId);
-    return { blocked };
-  }
-
-  @Get('projects/:id/dependency-graph')
-  async getDependencyGraph(@Param('id') projectId: string, @CurrentRequest() req: AppRequest) {
-    await this.domain.requireProjectRole(projectId, req.user.sub, ProjectRole.VIEWER);
-    return this.subtaskService.getDependencyGraph(projectId);
   }
 
   private toTaskCustomFieldFilterWhere(filter: TaskCustomFieldFilter): Prisma.TaskWhereInput {
